refactor(anthropic-proxy): USD & ModelID types, single pricing table, dedup TokenCounts
Phase 04: type money as USD and model identifiers as ModelID within the
anthropic proxy; replace the pricing map plus derived sorted-key slice
with one longest-prefix-first []pricingEntry table (no init sort, no sync
invariant); and collapse the three duplicate token-count declarations
into a single JSON-tagged TokenCounts embedded in sseInterceptor. JSON
wire, usage.json shape, and cost result text unchanged.

+++ internal/proxies/anthropic/proxy_test.go
@@ -257,5 +257,5 @@ data: {"type":"message_stop"}
 
 	snap := tracker.snapshot()
 	// 1k input @ $3/MTok + 2k output @ $15/MTok = $0.003 + $0.030 = $0.033.
-	assert.InDelta(t, 0.033, snap.CostUSD, 1e-9)
+	assert.InDelta(t, 0.033, float64(snap.CostUSD), 1e-9)
 }
